Move ContextFrame implementation back to header
The includes aren't so bad now that I (mostly) cleaned up EvaluationSession.h.
